Finish game fix
So here I had switched to a new branch dev to attempt to fix my broken game. Before (and is still the state of main currently) the game was trying to run through a loop. But, there was no great way to control the loop since the game progression depends on players clicking on the screen (as compared to a game where a player is prompted for a typed value which inherently has to wait for a response). So, instead of having a loop, I now have the logic of the game built into the click event. Once the game starts (which adds the events), each click validates the move, commits the move, checks for win/tie, then decides how to continue the game based on the move / board state.

In the past every game I've worked on had a game loop of some sort. It was interesting to build a game without a loop like this one (or rather, the game flow is controlled by the player actions rather than by a programmed loop).