Add button functionality and cell pointers, and change how cell validation works
Here I added functionality for the restart and new game buttons. I also added pointer cursors on each of the cells so its more responsive to a user. Finally I changed how move validation works. Instead of telling a user when a cell is taken, each time a cell is taken, the click event (and pointer cursor) is just removed from that cell. Now my program doesn't need to validate, the validation is just caked into the flow of the game.

At this point, the game works as intended! The code could use some refactoring certainly (but then again, can't it always). There are also some changes I could make to improve styling and functionality. Some of these are laid out in next_steps.txt. I'm not sure as of right now if I'm going to implement these changes on this project. Per usual, I try to get each project to just match the requirements, then add extras / refinement if I use it for a portfolio (or I suppose if I planned to publish it publicly).